refactor(frontend): vector layer colour maps
Replace the global `VECTOR_COLOR_MAPS` config with individual config for each vector layer eg. `config/networks/color-maps`, `config/regions/color-maps` etc.
